#123 modal clean up
css changes to modal.vue with supporting changes in modals for edit parameters, added to cag and import cag as needed to handle the new defaults as well as other minor clean up. Everything seems to look right with just the changes to the main modal as best I can tell.