#356 & #355 Insight Capture Clean Up
For #356 - layout fixes to new insight model to keep the image from breaking the scale of the capture.

For #355 - expanding the datacube-card capture area to include the header. doesn't yet put out anything from the dom as that can be very fragile, so should be scope as it's own work to add a class that can be applied to element we want to hide from captures that can be toggled hidden on capture, then turned back on save/cancel of catpure so we don't lose dom elements.