#416 Update Number Formatting In The Breakdown Panel
Replace the generic two decimal precision from the precision formatter with a specific d3.format setup, ',.2r' which will format any number with commas and provide a fixed two digits of precision when available in the original number. We may want to use similar formatting in other places where space isn't a concern, though for now I've kept to the area mentioned in the gitlab issue.