#531 reverting certain behaviors to fix the review checklist view
Went back to the getInsightByIds setup as that didn't seem to be the blocker anyway, and what was really critical was being able to refetch the insights after the questions had loaded. This also solves the linting issues introduced by trying to handle that mapping the setup. Still takes a second to pop in, but they pop in.